==> src/classifying_active_vs_controls_v2/01_make_10fold_dataset.ipynb <==
#01_make_10fold_dataset.py         Sep.5.2023
#Kevin

#objective: split the full dataset into 10fold
#caution [1]: class 0 and class 1 needs to be balanced.
#caution [2]: There should be 6 in the testing set and
#             54 in the training set

#input: "/Users/m197371/vasculitis_project_quantile_normalization/data/active_and_controls_machine_learning/active_and_controls_higher_in_active_machine_learning_all.csv"


import pandas as pd
# from sklearn.model_selection import StratifiedKFold
# from sklearn.model_selection import train_test_split
import os
import sys
sys.path.insert(1, '/Users/m197371/vasculitis_project_quantile_normalization/analysis/classifying_active_vs_control_v2')


data_df = pd.read_csv("../../data/active_and_controls_machine_learning/all_active_and_controls_machine_learning_final.csv", sep=",")

X_train_folds = []
X_test_folds = []
y_train_folds = []
y_test_folds = []

X = data_df.drop(columns=['Study_group','maskID','sample_ID','Classifier'])
y = data_df['Classifier']

num_samples, _ = data_df.shape

n_splits = 10

gca_idx_list = y[y == 1].index.tolist()
control_idx_list = y[y == 0].index.tolist()

samples_per_fold = (num_samples/n_splits)/2 #test set's total number of samples
kfold_index_dict = {}

index_count = 0
for i in range(n_splits):
    kfold = "%sfold" % (i + 1)

    for j in range(int(samples_per_fold)):
  
        gca_sample_idx = gca_idx_list[index_count]
        control_sample_idx = control_idx_list[index_count]

        try:
            kfold_index_dict[kfold].append(gca_sample_idx)
            kfold_index_dict[kfold].append(control_sample_idx)
        except KeyError:
            kfold_index_dict[kfold] = [gca_sample_idx, control_sample_idx]

        index_count += 1

#make index list of (1fold vs others) (2fold vs others) .... (10folds vs others)
for i in range(n_splits):
    kfold_i = "%sfold" % (i + 1) #this is for test dataset
    
    test_data_idx_list = kfold_index_dict[kfold_i]
    training_data_idx_list = []
    
    for j in range(n_splits):
        kfold_j= "%sfold" % (j + 1) #this is for training dataset
        
        if kfold_j != kfold_i:
            kfold_j_idx_list = kfold_index_dict[kfold_j]
            for temp_idx in kfold_j_idx_list:
                training_data_idx_list.append(temp_idx)
    
    print ("%s-fold dataset contains test (n = %s) and training (n = %s)" % (i+1, len(test_data_idx_list), len(training_data_idx_list)))
    set1 = set(test_data_idx_list)
    set2 = set(training_data_idx_list)

    intersection = set1.intersection(set2)
    print ("Common idxs: %s" % len(intersection))       
        
    #make 10fold dataset (.tsv or .csv format)
    test_data_df = data_df.loc[test_data_idx_list]
    training_data_df = data_df.loc[training_data_idx_list]
    
    folder_check = "../../analysis/classifying_active_vs_control_v2/%sfold" % (i + 1)
    if os.path.isdir(folder_check) == False:
        os.system('mkdir %s' % folder_check)
        
    test_data_file = "../../analysis/classifying_active_vs_control_v2/%sfold/test_data.csv" % (i+1)
    train_data_file = "../../analysis/classifying_active_vs_control_v2/%sfold/training_data.csv" % (i+1)
    
    test_data_df.to_csv(test_data_file)
    training_data_df.to_csv(train_data_file)
