==> src/classifying_active_vs_controls_v2/04_random_forest_classifier.ipynb <==
#05_random_forest_classifier.py         Sep.12.2023
#Kevin

#objective: read in each of the ten train and test sets: all proteins, the top proteins based on
#           fold-change*nominal pvalue value, and the top proteins based on the absolute value of 
#           fold-change*nominal pvalue value. Use the random forest classifier between GCA and 
#           Healthy Controls
#caution [1]: Be sure to use all the training and test sets and save them uniquely
#input: "/Users/m197371/vasculitis_project_quantile_normalization/analysis/classifying_active_vs_control_v2/"


from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import r2_score
import pandas as pd
import numpy as np

def main(working_dir):

    #classifying based on all features
    for i in range(1,11):
        #print("i: %s" % (i))
        train_file1 = "../../analysis/classifying_active_vs_control_v2/%sfold/train/%strain%s.csv" % (i,working_dir, i)
        #print(train_file1)
        train1 = pd.read_csv(train_file1)
        

        test_file1 = "../../analysis/classifying_active_vs_control_v2/%sfold/test/%stest%s.csv" % (i,working_dir, i)
        #print(test_file1)
        test1 = pd.read_csv(test_file1)
         
        feature_names = list(train1.columns.values)
        num_features = len(feature_names)
        feature_names = list(feature_names[4:num_features])
        
        x_train = train1[feature_names]
        y_train = train1['Classifier']

        x_test = test1[feature_names]
        y_test = test1['Classifier']
        rf_model = RandomForestClassifier(random_state=23)
        # rf_model = LogisticRegression()

        #print(feature_names)
        rf_model.fit(x_train, y_train)
        yhat1 = rf_model.predict(x_test)
        predicted_and_actual_values = pd.DataFrame()
        predicted_and_actual_values['Predicted'] = yhat1
        predicted_and_actual_values['Actual'] = y_test
        predicted_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/%spredicted_vals%s.csv" % (i,working_dir, i)
        predicted_and_actual_values.to_csv(predicted_file,index=False)   

        #features of importance
        features_of_importance = rf_model.feature_importances_
        features_of_importance = pd.DataFrame(features_of_importance)
        features_of_importance['Protein'] = feature_names
        #saving the dataframe
        #df.to_csv('file1.csv')
        features_of_importance_file = "../../analysis/classifying_active_vs_control_v2/%sfold/features_of_importance/%sfeatures_of_importance%s.csv" % (i,working_dir, i)
        features_of_importance.to_csv(features_of_importance_file,index=False)



temp_list = ['all_proteins_', 'top_10_proteins_', 'top_25_proteins_', 'top_50_proteins_',
             'top_100_proteins_', 'top_150_proteins_', 'top_200_proteins_','top_250_proteins_']

for working_dir in temp_list:
    main(working_dir)
    # break
print("done")
