==> src/classifying_active_vs_controls_v2/05_random_forest_results.ipynb <==
#07_differentially_abundant_and_absolute_value_proteins_random_forest_results.py      Sep.12.2023
#Kevin

#objective: read in each of the ten train and test sets results for the top proteins
#           based on the absolute value of fold-change*nominal pvalue value.

#caution [1]: Be sure to read in all the results one through ten
#input: "/Users/m197371/vasculitis_project_quantile_normalization/analysis/classifying_active_vs_control_v2/nfold/predicted/"


import pandas as pd
import numpy as np
def confusion_matrix_statistics_nfold(predicted_values):
    num_samples = len(predicted_values)
       
    true_positives = 0
    true_negatives = 0
    false_positives = 0
    false_negatives = 0
    for i in range(0,num_samples):
        predicted_val = predicted_values.iloc[i,0]
        actual_val = predicted_values.iloc[i,1]
        
        if(predicted_val == 1):
           if(actual_val == 1):
               true_positives += 1
           else:
               false_positives += 1

        if(predicted_val == 0):
           if(actual_val == 0):
               true_negatives += 1
           else:
               false_negatives += 1

    print("true positives: " + str(true_positives))
    print("true negatives: " + str(true_negatives))
    print("false positives: " + str(false_positives))
    print("false negatives: " + str(false_negatives))
    print("total number of samples: " + str(true_positives+true_negatives+false_positives+false_negatives))
    print("Accuracy: " + str((true_positives+true_negatives)/(true_positives+true_negatives+false_positives+false_negatives)))
    print("Specificity: " + str(true_negatives/(true_negatives + false_positives)))
    print("Sensitivity: " + str(true_positives/(true_positives + false_negatives)))
    print("PPV: " + str(true_positives/(true_positives + false_positives)))
    print("NPV: " + str(true_negatives/(true_negatives+false_negatives)))
    return

#read in results for all proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/all_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/all_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)

#read in results for the top 10 proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_10_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_10_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)

#read in results for the top 25 proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_25_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_25_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)

#read in results for the top 50 proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_50_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_50_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)

#read in results for the top 100 proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_100_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_100_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)

#read in results for the top 150 proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_150_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_150_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)

#read in results for the top 200 proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_200_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_200_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)

#read in results for the top 250 proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_250_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_250_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])

confusion_matrix_statistics_nfold(prediction_results1)
#print(prediction_results1)
