==> src/classifying_active_vs_controls_v2/06_differentially_abundant_proteins_random_forest_result.ipynb.ipynb <==
import pandas as pd
import numpy as np
def confusion_matrix_statistics_nfold(predicted_values):
    num_samples = len(predicted_values)
       
    true_positives = 0
    true_negatives = 0
    false_positives = 0
    false_negatives = 0
    for i in range(0,num_samples):
        predicted_val = predicted_values.iloc[i,0]
        actual_val = predicted_values.iloc[i,1]
        
        if(predicted_val == 1):
           if(actual_val == 1):
               true_positives += 1
           else:
               false_positives += 1

        if(predicted_val == 0):
           if(actual_val == 0):
               true_negatives += 1
           else:
               false_negatives += 1

    # print("true positives: " + str(true_positives))
    # print("true negatives: " + str(true_negatives))
    # print("false positives: " + str(false_positives))
    # print("false negatives: " + str(false_negatives))
    # print("total number of samples: " + str(true_positives+true_negatives+false_positives+false_negatives))
    # print("Accuracy: " + str((true_positives+true_negatives)/(true_positives+true_negatives+false_positives+false_negatives)))
    # print("Specificity: " + str(true_negatives/(true_negatives + false_positives)))
    # print("Sensitivity: " + str(true_positives/(true_positives + false_negatives)))
    # print("PPV: " + str(true_positives/(true_positives + false_positives)))
    # print("NPV: " + str(true_negatives/(true_negatives+false_negatives)))
    return [true_positives,true_negatives,false_positives, false_negatives,(true_positives+true_negatives+false_positives+false_negatives), ((true_positives+true_negatives)/(true_positives+true_negatives+false_positives+false_negatives)),(true_negatives/(true_negatives + false_positives)), (true_positives/(true_positives + false_negatives)), (true_positives/(true_positives + false_positives)), (true_negatives/(true_negatives+false_negatives))]   

#make a dataframe to hold the results, 7 different results
column_names = ['N_Proteins_included','true positives', 'true negatives', 'false positives', 'false negative','total number of samples', 'Accuracy', 'Specificity', 'Sensitivity', 'PPV', 'NPV']
all_random_forest_results = pd.DataFrame([[10, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]],columns=column_names)
print(all_random_forest_results)

#make a list of the protein cutoffs
#top_proteins_list = ["10","25","50","100","150","200","250"]
top_proteins_list = ["10"]

for cutoff in top_proteins_list:
    prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_%s_proteins_predicted_vals1.csv" % (cutoff) )

    for i in range(2,11):
        prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_%s_proteins_predicted_vals%s.csv" % (i, cutoff,i)
        prediction_results2 = pd.read_csv(prediction_file)
        prediction_results1 = pd.concat([prediction_results1, prediction_results2])
random_forest_results1 = (pd.DataFrame(confusion_matrix_statistics_nfold(prediction_results1)))
random_forest_results1 = pd.DataFrame(random_forest_results1.transpose())
#print(random_forest_results1.shape)
#print(random_forest_results1)
column_name = 'N_Proteins_included'
random_forest_results1.insert(0, column_name, 10)
random_forest_results1.columns = column_names
print(random_forest_results1)
all_random_forest_results = random_forest_results1

#now loop through the other results now that I have the results from the top 10 saved
top_proteins_list = ["25","50","100","150","200","250"]
#top_proteins_list = ["25"]

for cutoff in top_proteins_list:
    prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/top_%s_proteins_predicted_vals1.csv" % (cutoff) )
    #print("cutoff: %s" % cutoff)
    for i in range(2,11):
        prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/top_%s_proteins_predicted_vals%s.csv" % (i, cutoff,i)
        prediction_results2 = pd.read_csv(prediction_file)
        prediction_results1 = pd.concat([prediction_results1, prediction_results2])
    
    random_forest_results1 = (pd.DataFrame(confusion_matrix_statistics_nfold(prediction_results1)))
    random_forest_results1 = pd.DataFrame(random_forest_results1.transpose())
    #print(random_forest_results1.shape)
    #print(random_forest_results1)
    column_name = 'N_Proteins_included'
    random_forest_results1.insert(0, column_name, int(cutoff))
    random_forest_results1.columns = column_names
    #print(all_random_forest_results)
    all_random_forest_results =  pd.merge(all_random_forest_results,random_forest_results1, how='outer')
    #print(all_random_forest_results)
    

all_random_forest_results.to_csv("../../analysis/classifying_active_vs_control_v2/top_positive_protein_results.csv",index = False)
all_random_forest_results
#pd.concat([df1, df2], axis=1)
#pd.merge(df1, df2, on='key', how='inner')

#read in results for all proteins using 10-fold cross validation
#"all_proteins"

prediction_results1 = pd.read_csv("../../analysis/classifying_active_vs_control_v2/1fold/predicted/all_proteins_predicted_vals1.csv")
for i in range(2,11):
    prediction_file = "../../analysis/classifying_active_vs_control_v2/%sfold/predicted/all_proteins_predicted_vals%s.csv" % (i,i)
    prediction_results2 = pd.read_csv(prediction_file)
    prediction_results1 = pd.concat([prediction_results1, prediction_results2])
all_proteins_results = pd.DataFrame([[7289, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]],columns=column_names)
all_proteins_results_to_add = confusion_matrix_statistics_nfold(prediction_results1)
print(all_proteins_results_to_add)
all_proteins_results.at[0, 'true positives'] = all_proteins_results_to_add[0]
all_proteins_results.at[0, 'true negatives'] = all_proteins_results_to_add[1]
all_proteins_results.at[0, 'false positives'] = all_proteins_results_to_add[2]
all_proteins_results.at[0, 'false negative'] = all_proteins_results_to_add[3]
all_proteins_results.at[0, 'total number of samples'] = all_proteins_results_to_add[4]
all_proteins_results.at[0, 'Accuracy'] = all_proteins_results_to_add[5]
all_proteins_results.at[0, 'Specificity'] = all_proteins_results_to_add[6]
all_proteins_results.at[0, 'Sensitivity'] = all_proteins_results_to_add[7]
all_proteins_results.at[0, 'PPV'] = all_proteins_results_to_add[8]
all_proteins_results.at[0, 'NPV'] = all_proteins_results_to_add[9]
print(all_proteins_results)
#print(prediction_results1)

#now add the results to the last row of the all_random_forest_results
all_random_forest_results =  pd.merge(all_random_forest_results,all_proteins_results, how='outer')
all_random_forest_results

print(all_random_forest_results)
all_random_forest_results.to_csv("../../analysis/classifying_active_vs_control_v2/top_positive_protein_results.csv",index = False)
